+++ pkg/sass_api/CHANGELOG.md
@@ -1,6 +1,12 @@
-## 14.5.0
+## 15.0.0
 
-* No user-visible changes.
+* Rename `ArgumentInvocation` to `ArgumentList`, `ArgumentDeclaration` to
+  `ParameterList`, and `Argument` to `Parameter` to better match the
+  conventional distinction between "arugments" and "parameters".
+
+* Rename `ArgumentDeclaration.arguments` to `ParameterList.parameters`,
+  `ArgumentDeclaration.restArgument` to `ParameterList.restParameter`,
+  `CallableDeclaration.arguments` to `.parameters`.
 
 ## 14.4.1-dev